Interval in tracker_response parsed

+++ src/protocol.h
@@ -14,7 +14,7 @@ typedef struct
 
 typedef struct
 {
-  int interval;
+  long interval;
   peer *peers;
 } tracker_response;
